spark_processing/requirements.txt
pyspark
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
kafka-python